+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - `AvatarInitials`: set `letter-spacing` to zero to center the text. ([@TaraldRotsaert](https://github.com/TaraldRotsaert) in [#575](https://github.com/teamleadercrm/ui/pull/575))
+- `AvatarStack`: remove `size` prop of `AvatarStack` wrapper. ([@TaraldRotsaert](https://github.com/TaraldRotsaert) in [#576](https://github.com/teamleadercrm/ui/pull/576))
 
 ## [0.24.4] - 2019-03-26